update bagian backend

+++ app/Http/Controllers/Admin/SettingController.php
@@ -32,7 +32,10 @@ class SettingController extends Controller
 
         if(request()->file('favicon'))
         {
-            Storage::disk('public')->delete($setting->favicon);
+            if($setting->favicon)
+            {
+                Storage::disk('public')->delete($setting->favicon);
+            }
             $data['favicon'] = request()->file('favicon')->store('settings','public');
         }else{
             $data['favicon'] = $setting->favicon;
@@ -40,7 +43,10 @@ class SettingController extends Controller
 
         if(request()->file('image'))
         {
-            Storage::disk('public')->delete($setting->image);
+            if($setting->image)
+            {
+                Storage::disk('public')->delete($setting->image);
+            }
             $data['image'] = request()->file('image')->store('settings','public');
         }else{
             $data['image'] = $setting->image;
